Change cpp_bin_float::convert_to to round correctly to float. Also add another test case from https://svn.boost.org/trac/boost/ticket/12039, plus test case to check rounding of ties.

+++ include/boost/multiprecision/cpp_int/bitwise.hpp
@@ -325,7 +325,7 @@ inline void left_shift_byte(Int& result, double_limb_type s)
 
    if(rs != ors)
       pr[rs - 1] = 0u;
-   std::size_t bytes = s / CHAR_BIT;
+   std::size_t bytes = static_cast<std::size_t>(s / CHAR_BIT);
    std::size_t len = std::min(ors * sizeof(limb_type), rs * sizeof(limb_type) - bytes);
    if(bytes >= rs * sizeof(limb_type))
       result = static_cast<limb_type>(0u);
